refactor: split generator into submodules, rename date fns, extract RetryWriter
Generator refactor (ffwd-io):
- Split monolithic generator.rs into focused submodules using include!
- Rename date math functions to compute_/convert_ prefix convention:
  - days_from_civil -> compute_days_from_civil
  - civil_from_days -> compute_civil_from_days
  - epoch_ms_to_parts -> convert_epoch_ms_to_parts
- Extract platform utilities from schema.rs into platform.rs

Output refactor (ffwd-output):
- Extract RetryWriter combinator for at-least-once write semantics
- 5 scripted test scenarios covering retry behavior
