+++ crates/ffwd-io/src/host_metrics/platform.rs
@@ -1,0 +1,57 @@
+// xtask-verify: allow(pub_module_needs_tests) // trivial wrappers over OS APIs; behaviour is compile-time platform detection
+
+/// The host operating system family detected at compile time.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum HostPlatform {
+    Linux,
+    Macos,
+    Windows,
+    Unsupported,
+}
+
+impl HostPlatform {
+    /// Returns the lowercase platform name, or `None` for unsupported platforms.
+    pub const fn as_str(self) -> Option<&'static str> {
+        match self {
+            Self::Linux => Some("linux"),
+            Self::Macos => Some("macos"),
+            Self::Windows => Some("windows"),
+            Self::Unsupported => None,
+        }
+    }
+}
+
+/// Convert an `OsStr` to a `String`, replacing non-UTF-8 sequences with the replacement character.
+pub fn os_to_string(value: &OsStr) -> String {
+    value.to_string_lossy().to_string()
+}
+
+/// Convert a slice of `OsString` values to a single space-separated `String`.
+pub fn os_vec_to_string(values: &[std::ffi::OsString]) -> String {
+    values
+        .iter()
+        .map(|v| v.to_string_lossy())
+        .collect::<Vec<_>>()
+        .join(" ")
+}
+
+/// Return the current wall-clock time as nanoseconds since the Unix epoch.
+pub fn now_unix_nano() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_nanos() as u64
+}
+
+/// Detect the host operating system at compile time and return the corresponding [`HostPlatform`].
+pub fn current_host_platform() -> HostPlatform {
+    if cfg!(target_os = "linux") {
+        HostPlatform::Linux
+    } else if cfg!(target_os = "macos") {
+        HostPlatform::Macos
+    } else if cfg!(target_os = "windows") {
+        HostPlatform::Windows
+    } else {
+        HostPlatform::Unsupported
+    }
+}
