+++ crates/ffwd-io/src/host_metrics/schema.rs
@@ -9,53 +9,3 @@ fn sensor_schema() -> Arc<Schema> {
 
     sensor_row_columns!(schema_from_sensor_columns)
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum HostPlatform {
-    Linux,
-    Macos,
-    Windows,
-    Unsupported,
-}
-
-impl HostPlatform {
-    const fn as_str(self) -> Option<&'static str> {
-        match self {
-            Self::Linux => Some("linux"),
-            Self::Macos => Some("macos"),
-            Self::Windows => Some("windows"),
-            Self::Unsupported => None,
-        }
-    }
-}
-
-fn os_to_string(value: &OsStr) -> String {
-    value.to_string_lossy().to_string()
-}
-
-fn os_vec_to_string(values: &[std::ffi::OsString]) -> String {
-    values
-        .iter()
-        .map(|v| v.to_string_lossy())
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn now_unix_nano() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_nanos() as u64
-}
-
-fn current_host_platform() -> HostPlatform {
-    if cfg!(target_os = "linux") {
-        HostPlatform::Linux
-    } else if cfg!(target_os = "macos") {
-        HostPlatform::Macos
-    } else if cfg!(target_os = "windows") {
-        HostPlatform::Windows
-    } else {
-        HostPlatform::Unsupported
-    }
-}
